+++ public/js/controller/displayCdhController.js
@@ -1,9 +1,23 @@
 spaceFrontierApp.controller("displayCdhController", function($scope, $http) {
   	$scope.parts = [];
     $scope.newPart = {};
-    $scope.editNewPart = {};
+    $scope.notAdmin = false;
 
     var token = window.location.search.slice(1);
+
+    $scope.remove = function() {
+        $scope.notAdmin = false;
+    }
+
+    $scope.decode = function() {
+        var payload = token.split('=')[1];
+        var base64Url = payload.split('.')[1];
+        var output = JSON.parse(atob(base64Url));
+        $scope.name = output.name;
+        $scope.email = output.email;
+        $scope.admin = output.admin;
+    };
+    $scope.decode();
 
     $scope.backMain = function() {
         var newPath = '/listAdmin?' + token;
@@ -19,6 +33,7 @@ spaceFrontierApp.controller("displayCdhController", function($scope, $http) {
 
     $scope.predicate = 'Manufacturer';
     $scope.reverse = false;
+
     $scope.order = function(predicate) {
         $scope.reverse = ($scope.predicate === predicate) ? !$scope.reverse : false;
         $scope.predicate = predicate;
@@ -64,29 +79,51 @@ spaceFrontierApp.controller("displayCdhController", function($scope, $http) {
             Additional_info: $scope.info, 
             Cost: "0"
         };
-        $http.put('/parts/cdh/' + $scope.id + '?' + token, $scope.newPart)
-            .success(function(data) {
-                $scope.id = '';
-                $scope.type = ''; 
-                $scope.name = ''; 
-                $scope.manu = ''; 
-                $scope.ref = '';
-                $scope.her = '';
-                $scope.mass = '';
-                $scope.power = '';
-                $scope.vol = '';
-                $scope.prop = '';
-                $scope.mips = '';
-                $scope.mhz = '';
-                $scope.process = '';
-                $scope.mem = '';
-                $scope.sci = '';
-                $scope.info = '';
-                $http.get('/parts/cdh').then(function(result) { 
-                    $scope.parts = result.data; 
-                });
-        });
-        $scope.edit = false;
+        if ($scope.admin) {
+            $http.put('/parts/cdh/' + $scope.id + '?' + token, $scope.newPart)
+                .success(function(data) {
+                    $scope.id = '';
+                    $scope.type = ''; 
+                    $scope.name = ''; 
+                    $scope.manu = ''; 
+                    $scope.ref = '';
+                    $scope.her = '';
+                    $scope.mass = '';
+                    $scope.power = '';
+                    $scope.vol = '';
+                    $scope.prop = '';
+                    $scope.mips = '';
+                    $scope.mhz = '';
+                    $scope.process = '';
+                    $scope.mem = '';
+                    $scope.sci = '';
+                    $scope.info = '';
+                    $http.get('/parts/cdh').then(function(result) { 
+                        $scope.parts = result.data; 
+                    });
+                    $scope.edit = false;            
+            });
+        } else {
+            $scope.notAdmin = true;
+            $scope.message = 'You do not have permission to edit this item';
+            $scope.id = '';
+            $scope.type = ''; 
+            $scope.name = ''; 
+            $scope.manu = ''; 
+            $scope.ref = '';
+            $scope.her = '';
+            $scope.mass = '';
+            $scope.power = '';
+            $scope.vol = '';
+            $scope.prop = '';
+            $scope.mips = '';
+            $scope.mhz = '';
+            $scope.process = '';
+            $scope.mem = '';
+            $scope.sci = '';
+            $scope.info = '';        
+            $scope.edit = false;            
+        }    
     }
 
     $scope.openEdit = function(part) {
@@ -156,12 +193,17 @@ spaceFrontierApp.controller("displayCdhController", function($scope, $http) {
     };
 
     $scope.deletePart = function(id) {
-        $http.delete('/parts/cdh/' + id + '?' + token)
-            .success(function(data) {
-                $http.get('/parts/cdh').then(function(result) { 
-                    $scope.parts = result.data; 
-                });
+        if ($scope.admin) {
+            $http.delete('/parts/cdh/' + id + '?' + token)
+                .success(function(data) {
+                    $http.get('/parts/cdh').then(function(result) { 
+                        $scope.parts = result.data; 
+                    });
             });
+        } else {
+            $scope.notAdmin = true;
+            $scope.message = 'You do not have permission to delete this item';
+        }
     };
 
     $('#exampleModal').on('show.bs.modal', function (event) {
