+++ public/js/controller/displayStationController.js
@@ -1,8 +1,23 @@
 spaceFrontierApp.controller("displayStationController", function($scope, $http) {
     $scope.parts = [];
     $scope.newPart = {};
+    $scope.notAdmin = false;
     
     var token = window.location.search.slice(1);
+
+    $scope.remove = function() {
+        $scope.notAdmin = false;
+    }
+
+    $scope.decode = function() {
+        var payload = token.split('=')[1];
+        var base64Url = payload.split('.')[1];
+        var output = JSON.parse(atob(base64Url));
+        $scope.name = output.name;
+        $scope.email = output.email;
+        $scope.admin = output.admin;
+    };
+    $scope.decode();
 
     $scope.backMain = function() {
         var newPath = '/listAdmin?' + token;
@@ -67,23 +82,39 @@ spaceFrontierApp.controller("displayStationController", function($scope, $http) 
             Loc: $scope.loc, 
             Band: $scope.band,
         };
-        $http.put('/parts/station/' + $scope.id + '?' + token, $scope.newPart)
-            .success(function(data) {
-                $scope.id = '';
-                $scope.name = '';
-                $scope.uplink = '';
-                $scope.eirp = '';
-                $scope.down = '';
-                $scope.gain = '';
-                $scope.pgain = '';
-                $scope.dia = '';
-                $scope.loc = '';
-                $scope.band = '';
-                $http.get('/parts/station').then(function(result) { 
-                    $scope.parts = result.data; 
-                });
-        });        
-        $scope.edit = false;
+        if ($scope.admin) {
+            $http.put('/parts/station/' + $scope.id + '?' + token, $scope.newPart)
+                .success(function(data) {
+                    $scope.id = '';
+                    $scope.name = '';
+                    $scope.uplink = '';
+                    $scope.eirp = '';
+                    $scope.down = '';
+                    $scope.gain = '';
+                    $scope.pgain = '';
+                    $scope.dia = '';
+                    $scope.loc = '';
+                    $scope.band = '';
+                    $http.get('/parts/station').then(function(result) { 
+                        $scope.parts = result.data; 
+                    });
+                    $scope.edit = false;
+            });     
+        } else {
+            $scope.notAdmin = true;
+            $scope.message = 'You do not have permission to edit this item';
+            $scope.id = '';
+            $scope.name = '';
+            $scope.uplink = '';
+            $scope.eirp = '';
+            $scope.down = '';
+            $scope.gain = '';
+            $scope.pgain = '';
+            $scope.dia = '';
+            $scope.loc = '';
+            $scope.band = '';
+            $scope.edit = false;
+        }   
     }
 
     $scope.addPart = function() {
@@ -117,12 +148,17 @@ spaceFrontierApp.controller("displayStationController", function($scope, $http) 
     };
 
     $scope.deletePart = function(id) {
-        $http.delete('/parts/station/' + id + '?' + token)
-            .success(function(data) {
-                $http.get('/parts/station').then(function(result) { 
-                    $scope.parts = result.data; 
-                });
+        if ($scope.admin) {
+            $http.delete('/parts/station/' + id + '?' + token)
+                .success(function(data) {
+                    $http.get('/parts/station').then(function(result) { 
+                        $scope.parts = result.data; 
+                    });
             });
+        } else {
+            $scope.notAdmin = true;
+            $scope.message = 'You do not have permission to delete this item';
+        }
     };
 
     $('#exampleModal').on('show.bs.modal', function (event) {
